refactor: Renomeado a Solution, Projeto e Pastas atodando o PascalCase, removendo os espaços entre Palavras
