--- Sistema de Estacionamento/DataBase/IEF - Interface/IExecution_ef.cs
@@ -1,25 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using static Microsoft.EntityFrameworkCore.DbLoggerCategory;
-
-namespace Sistema_de_Estacionamento.DataBase.IEF___Interface
-{
-    internal interface IExecution_ef
-    {
-
-        void Insert_EF() { }
-        void Insert_CheckOut() { }
-        void Delete_EF() { }
-        bool ValidacaoCredencial_EF(string Credencial) { return true; }
-        void QueryCredential_EF(string Credencial) { }
-        void QueryPlate_EF(string Plate) { }
-
-        void Query_All() { }
-        void Query_parked() { }
-        void Query_specific() { }
-
-    }
-}
