+++ SistemaDeEstacionamento.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.11.35312.102
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SistemaDeEstacionamento", "SistemaDeEstacionamento\SistemaDeEstacionamento.csproj", "{080FD7C1-3886-4355-9D35-F4AC07B9DA74}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{080FD7C1-3886-4355-9D35-F4AC07B9DA74}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{080FD7C1-3886-4355-9D35-F4AC07B9DA74}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{080FD7C1-3886-4355-9D35-F4AC07B9DA74}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{080FD7C1-3886-4355-9D35-F4AC07B9DA74}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F875E006-FDF2-4A4D-BE44-7375C7A5DA68}
+	EndGlobalSection
+EndGlobal
